refactor: Extract test utils
