+++ packages/cozy-konnector-libs/src/libs/testUtils.js
@@ -96,7 +96,29 @@ const wrapAsFetchJSONResult = documents => {
   }
 }
 
+// Uses before/after each to mock env variables during tests
+const mockEnvVariables = attrs => {
+  let original = {}
+  beforeEach(() => {
+    for (let k in attrs) {
+      original[k] = process.env[k]
+      process.env[k] = attrs[k]
+    }
+  })
+
+  afterEach(() => {
+    for (let k in attrs) {
+      process.env[k] = original[k]
+    }
+  })
+}
+
+const asyncResolve = data =>
+  new Promise(resolve => setImmediate(() => resolve(data)))
+
 module.exports = {
+  mockEnvVariables,
+  asyncResolve,
   mkLineParser,
   parseTable,
   wrapAsFetchJSONResult
